Create handle_choices.rb file. Move the handle_choce method from main.rb to the new file.
